agent/adk/quick_start/adk-streaming/pyproject.toml
[project]
name = "adk-streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.2.1",
]
